each user can see his own categories

+++ lib/utils/mongo/categories.ts
@@ -4,12 +4,12 @@ import mongoose from "mongoose";
 
 mongoose.connect(process.env.MONGO_URI!).then(() => console.log("connected"));
 
-export async function getCategoryId(name: String) {
-  return Category.findOne({ name: name });
+export async function getCategoryId(name: String, user: string) {
+  return Category.findOne({ name: name, owner: user });
 }
 
-export async function addCategory(name: String) {
-  const newCat = new Category({ name: name });
+export async function addCategory(name: String, user: string) {
+  const newCat = new Category({ name: name, owner: user });
   return newCat.save();
 }
 
@@ -17,13 +17,20 @@ export async function getCategoryName(id: string) {
   return Category.findOne({ _id: id });
 }
 
-export async function showAllCategories() {
-  return Category.find({});
+export async function showAllCategories(user: string) {
+  return Category.find({ owner: user });
 }
 
-export async function editCategory(oldName: String, newName: String) {
-  return Category.updateOne({ name: oldName }, { name: newName });
+export async function editCategory(
+  oldName: String,
+  newName: String,
+  user: string
+) {
+  return Category.updateOne(
+    { owner: user, name: oldName },
+    { owner: user, name: newName }
+  );
 }
-export async function deleteCategory(name: String) {
-  return Category.deleteOne({ name: name });
+export async function deleteCategory(name: string, user: string) {
+  return Category.deleteOne({ name: name, owner: user });
 }
